# layout.completely_flatten should not concatenate (performance issue).