--- a/src/awkward/_v2/operations/reducers/ak_any.py
+++ b/src/awkward/_v2/operations/reducers/ak_any.py
@@ -44,10 +44,20 @@ def any(array, axis=None, keepdims=False, mask_identity=False, flatten_records=F
     )
 
     if axis is None:
-        return layout.nplike.any(
-            layout.completely_flatten(
-                function_name="ak.any", flatten_records=flatten_records
-            )
+
+        def reduce(xs):
+            if len(xs) == 1:
+                return xs[0]
+            else:
+                return layout.nplike.logical_or(xs[0], reduce(xs[1:]))
+
+        return reduce(
+            [
+                layout.nplike.any(x)
+                for x in layout.completely_flatten(
+                    function_name="ak.any", flatten_records=flatten_records
+                )
+            ]
         )
 
     else: