--- a/src/awkward/_v2/operations/reducers/ak_argmax.py
+++ b/src/awkward/_v2/operations/reducers/ak_argmax.py
@@ -52,13 +52,18 @@ def argmax(array, axis=None, keepdims=False, mask_identity=True, flatten_records
         layout = ak._v2.operations.structure.fill_none(
             layout, -np.inf, axis=-1, highlevel=False
         )
-        flat = layout.completely_flatten(
+
+        best_index = None
+        best_value = None
+        for tmp in layout.completely_flatten(
             function_name="ak.argmax", flatten_records=flatten_records
-        )
-        if len(flat) == 0:
-            return None
-        else:
-            return layout.nplike.argmax(flat)
+        ):
+            # FIXME: this isn't going to survive a type-tracer!
+            out = layout.nplike.argmax(tmp, axis=None)
+            if best_index is None or tmp[out] > best_value:
+                best_index = out
+                best_value = tmp[out]
+        return best_index
 
     else:
         behavior = ak._v2._util.behavior_of(array)