--- a/src/awkward/_v2/operations/reducers/ak_count.py
+++ b/src/awkward/_v2/operations/reducers/ak_count.py
@@ -83,10 +83,20 @@ def count(array, axis=None, keepdims=False, mask_identity=False, flatten_records
     )
 
     if axis is None:
-        return layout.nplike.size(
-            layout.completely_flatten(
-                function_name="ak.count", flatten_records=flatten_records
-            )
+
+        def reduce(xs):
+            if len(xs) == 1:
+                return xs[0]
+            else:
+                return layout.nplike.add(xs[0], reduce(xs[1:]))
+
+        return reduce(
+            [
+                layout.nplike.size(x)
+                for x in layout.completely_flatten(
+                    function_name="ak.count", flatten_records=flatten_records
+                )
+            ]
         )
 
     else: