--- a/src/awkward/_v2/operations/reducers/ak_min.py
+++ b/src/awkward/_v2/operations/reducers/ak_min.py
@@ -56,13 +56,19 @@ def min(
     )
 
     if axis is None:
-        flat = layout.completely_flatten(
+
+        def reduce(xs):
+            if len(xs) == 0:
+                return None
+            elif len(xs) == 1:
+                return xs[0]
+            else:
+                return layout.nplike.minimum(xs[0], reduce(xs[1:]))
+
+        tmp = layout.completely_flatten(
             function_name="ak.min", flatten_records=flatten_records
         )
-        if len(flat) == 0:
-            return None if mask_identity else np.iinfo(flat.dtype.type).max
-        else:
-            return layout.nplike.min(flat)
+        return reduce([layout.nplike.min(x) for x in tmp if len(x) > 0])
 
     else:
         behavior = ak._v2._util.behavior_of(array)