--- a/src/awkward/_v2/operations/reducers/ak_prod.py
+++ b/src/awkward/_v2/operations/reducers/ak_prod.py
@@ -42,10 +42,20 @@ def prod(array, axis=None, keepdims=False, mask_identity=False, flatten_records=
     )
 
     if axis is None:
-        return layout.nplike.prod(
-            layout.completely_flatten(
-                function_name="ak.prod", flatten_records=flatten_records
-            )
+
+        def reduce(xs):
+            if len(xs) == 1:
+                return xs[0]
+            else:
+                return layout.nplike.multiply(xs[0], reduce(xs[1:]))
+
+        return reduce(
+            [
+                layout.nplike.prod(x)
+                for x in layout.completely_flatten(
+                    function_name="ak.prod", flatten_records=flatten_records
+                )
+            ]
         )
 
     else: