--- a/src/awkward/_v2/operations/structure/ak_flatten.py
+++ b/src/awkward/_v2/operations/structure/ak_flatten.py
@@ -99,7 +99,7 @@ def flatten(array, axis=1, highlevel=True, behavior=None):
     nplike = ak.nplike.of(layout)
 
     if axis is None:
-        out = (layout.completely_flatten(),)
+        out = layout.completely_flatten(function_name="ak.flatten")
         assert isinstance(out, tuple) and all(isinstance(x, np.ndarray) for x in out)
 
         out = ak._v2.contents.NumpyArray(nplike.concatenate(out))