--- a/src/awkward/_v2/operations/structure/ak_ravel.py
+++ b/src/awkward/_v2/operations/structure/ak_ravel.py
@@ -49,7 +49,7 @@ def ravel(array, highlevel=True, behavior=None):
     )
     nplike = ak.nplike.of(layout)
 
-    out = (layout.completely_flatten(),)
+    out = layout.completely_flatten(function_name="ak.ravel")
     assert isinstance(out, tuple) and all(isinstance(x, np.ndarray) for x in out)
 
     if any(isinstance(x, nplike.ma.MaskedArray) for x in out):