This was raised on Slack by @martindurant and @douglasdavis.

**Before:**

```python
In [1]: import awkward._v2 as ak, numpy as np

In [2]: x = np.ones((100, 100, 100))

In [3]: arr = ak.Array(x)

In [4]: %timeit ak.sum(arr, axis=None)
813 µs ± 6.08 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [5]: %timeit x.sum()
332 µs ± 5.38 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

**After:**

```python
In [1]: import awkward._v2 as ak, numpy as np

In [2]: x = np.ones((100, 100, 100))

In [3]: arr = ak.Array(x)

In [4]: %timeit ak.sum(arr, axis=None)
317 µs ± 20.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [5]: %timeit x.sum()
295 µs ± 8.65 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

It removes an unnecessary concatenation.